recursos solicitan token de autenticacion

+++ resources/evento.py
@@ -1,11 +1,16 @@
 from flask import Flask, Blueprint, jsonify, request
 from models.evento import Evento
 from models.curso import Curso
+from models.administrador import Administrador
+from models.alumno import Alumno
+from models.apoderado import Apoderado
+from models.profesor import Profesor
 from flask_restful import Api, Resource, url_for
 from libs.to_dict import mongo_to_dict
 from datetime import datetime
 import json
 from bson import json_util
+from flask_restful import reqparse
 
 def init_module(api):
     api.add_resource(EventoItem, '/eventos/<id>')
@@ -13,15 +18,43 @@ def init_module(api):
     api.add_resource(EventosSolicitudes, '/eventos/solicitudes')
 
 class EventoItem(Resource):
+    def __init__(self):
+        self.reqparse = reqparse.RequestParser()
+        self.reqparse.add_argument('auth-token', type = str, required=True, location='headers')
+        super(EventoItem, self).__init__()
     def get(self, id):
-        return json.loads(Colegio.objects(id=id).first().to_json())
+        args = self.reqparse.parse_args()
+        token = args.get('auth-token')
+        alumno = Alumno.load_from_token(token)
+        apoderado = Apoderado.load_from_token(token)
+        administrador = Administrador.load_from_token(token)
+        profesor = Profesor.load_from_token(token)
+        if alumno == None and apoderado == None and administrador == None and profesor == None:
+            return {'response': 'user_invalid'},401
+        return json.loads(Evento.objects(id=id).first().to_json())
     
     def delete(self,id):
+        args = self.reqparse.parse_args()
+        token = args.get('auth-token')
+        alumno = Alumno.load_from_token(token)
+        apoderado = Apoderado.load_from_token(token)
+        administrador = Administrador.load_from_token(token)
+        profesor = Profesor.load_from_token(token)
+        if alumno == None and apoderado == None and administrador == None and profesor == None:
+            return {'response': 'user_invalid'},401
         evento = Evento.objects(id=id).first()
         evento.delete()
         return {'Response':'exito'}
     
     def put(self,id):
+        args = self.reqparse.parse_args()
+        token = args.get('auth-token')
+        alumno = Alumno.load_from_token(token)
+        apoderado = Apoderado.load_from_token(token)
+        administrador = Administrador.load_from_token(token)
+        profesor = Profesor.load_from_token(token)
+        if alumno == None and apoderado == None and administrador == None and profesor == None:
+            return {'response': 'user_invalid'},401
         evento = Evento.objects(id=id).first()
         evento.activo = True
         evento.save()
@@ -29,7 +62,19 @@ class EventoItem(Resource):
 
 
 class Eventos(Resource):
+    def __init__(self):
+        self.reqparse = reqparse.RequestParser()
+        self.reqparse.add_argument('auth-token', type = str, required=True, location='headers')
+        super(Eventos, self).__init__()
     def get(self):
+        args = self.reqparse.parse_args()
+        token = args.get('auth-token')
+        alumno = Alumno.load_from_token(token)
+        apoderado = Apoderado.load_from_token(token)
+        administrador = Administrador.load_from_token(token)
+        profesor = Profesor.load_from_token(token)
+        if alumno == None and apoderado == None and administrador == None and profesor == None:
+            return {'response': 'user_invalid'},401
         response = []
         eventos = Evento.objects().all()
         for evento in eventos:
@@ -38,6 +83,14 @@ class Eventos(Resource):
         return response
     
     def post(self):
+        args = self.reqparse.parse_args()
+        token = args.get('auth-token')
+        alumno = Alumno.load_from_token(token)
+        apoderado = Apoderado.load_from_token(token)
+        administrador = Administrador.load_from_token(token)
+        profesor = Profesor.load_from_token(token)
+        if alumno == None and apoderado == None and administrador == None and profesor == None:
+            return {'response': 'user_invalid'},401
         data = request.data.decode()
         data = json.loads(data)
         evento = Evento()
@@ -50,7 +103,19 @@ class Eventos(Resource):
         return {'Response':'exito'}
 
 class EventosSolicitudes(Resource):
+    def __init__(self):
+        self.reqparse = reqparse.RequestParser()
+        self.reqparse.add_argument('auth-token', type = str, required=True, location='headers')
+        super(EventosSolicitudes, self).__init__()
     def get(self):
+        args = self.reqparse.parse_args()
+        token = args.get('auth-token')
+        alumno = Alumno.load_from_token(token)
+        apoderado = Apoderado.load_from_token(token)
+        administrador = Administrador.load_from_token(token)
+        profesor = Profesor.load_from_token(token)
+        if alumno == None and apoderado == None and administrador == None and profesor == None:
+            return {'response': 'user_invalid'},401
         response = []
         eventos = Evento.objects().all()
         for evento in eventos:
